Revert "add summary memory stream agent"
This reverts commit dc380bfec5bf91861016d56abf11cd49868d86d2.
