+++ doc/prompt/memory_stream/common_prompt.json
@@ -3,7 +3,6 @@
         "現在正在進行狼人殺的遊戲，請你根據下面的例子及提供的資訊，依照格式進行回答，深入思考並猜測該號玩家是什麼身分，以下是一些例子:",
         "%e",
         "請參考以上例子，給予以下玩家資訊，猜測[該玩家的身分]，只允許回答「好人、神職、狼人、女巫、預言家、村民、獵人」",
-        "%s",
         "[目標]",
         "%t",
         "[該玩家的資訊]",
@@ -57,7 +56,6 @@
         "%e",
         "給你以下資訊，請根據上述的例子，仔細思考今天應該要票誰出去，並且依照格式進行輸出。",
         "目前可以進行[投票]的對象 : 「%t、棄票」",
-        "%s",
         "* 資訊",
         "[玩家身分猜測]",
         "%l",
@@ -69,8 +67,6 @@
     "dialogue" : [
         "現在正在進行狼人殺遊戲，你正要準備發言，請參考以下例子，決定你的發言:",
         "%e",
-        "",
-        "%s",
         "*資訊",
         "[玩家身分猜測]",
         "%l",
