--- doc/prompt/memory_stream/summary_addition_prompt.json
@@ -1,5 +1,0 @@
-{
-    "summary" : [
-        "以下是提供給您的一些遊玩建議:"
-    ]
-}
